+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## Unreleased
 
 - Add `multi` to Output Objects, as a way to define more than one field of the same type at once.
+- Add an `inherits:` key to output objects, for view inheritance.